requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Image processing
Pillow==10.1.0
pytesseract==0.3.10

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# CORS middleware
python-cors==1.7.0

# Optional: For enhanced OCR (Korean language support)
# easyocr==1.7.0

# Optional: For advanced image preprocessing
# opencv-python==4.8.1.78
# numpy==1.24.3

# Development dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1